requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.9

# Security
bcrypt==4.1.2

# Validation
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.1

# File Storage
minio==7.2.0

# Utilities
Werkzeug==3.0.1

# Optional: For better development
python-decouple==3.8